+++ rendering/lsystem/segment.py
